+++ deal.II/contrib/umfpack/UMFPACK/Source/umf_init_front.c
@@ -65,8 +65,8 @@ GLOBAL Int UMF_init_front
 #endif
     if (Work->do_grow)
     {
-	fnr2 = UMF_FRONTAL_GROWTH * Work->fnrows_new + 2 ;
-	fnc2 = UMF_FRONTAL_GROWTH * Work->fncols_new + 2 ;
+	fnr2 = (Int) (UMF_FRONTAL_GROWTH * Work->fnrows_new + 2) ;
+	fnc2 = (Int) (UMF_FRONTAL_GROWTH * Work->fncols_new + 2) ;
 	if (!UMF_grow_front (Numeric, fnr2, fnc2, Work,
 	    Work->pivrow_in_front ? 2 : 0))
 	{
