+++ deal.II/contrib/umfpack/UMFPACK/Source/umfpack_report_control.c
@@ -25,7 +25,7 @@ GLOBAL void UMFPACK_report_control
     double drow, dcol, relpt, relpt2, alloc_init, front_alloc_init, amd_alpha,
 	tol, force_fixQ, droptol, aggr ;
 
-    prl = GET_CONTROL (UMFPACK_PRL, UMFPACK_DEFAULT_PRL) ;
+    prl = (Int) GET_CONTROL (UMFPACK_PRL, UMFPACK_DEFAULT_PRL) ;
 
     if (prl < 2)
     {
@@ -96,7 +96,7 @@ GLOBAL void UMFPACK_report_control
     /* block size */
     /* ---------------------------------------------------------------------- */
 
-    nb = GET_CONTROL (UMFPACK_BLOCK_SIZE, UMFPACK_DEFAULT_BLOCK_SIZE) ;
+    nb = (Int) GET_CONTROL (UMFPACK_BLOCK_SIZE, UMFPACK_DEFAULT_BLOCK_SIZE) ;
     nb = MAX (1, nb) ;
     PRINTF (("    "ID": block size for dense matrix kernels: "ID"\n",
 	(Int) INDEX (UMFPACK_BLOCK_SIZE), nb)) ;
@@ -105,7 +105,7 @@ GLOBAL void UMFPACK_report_control
     /* strategy */
     /* ---------------------------------------------------------------------- */
 
-    strategy = GET_CONTROL (UMFPACK_STRATEGY, UMFPACK_DEFAULT_STRATEGY) ;
+    strategy = (Int) GET_CONTROL (UMFPACK_STRATEGY, UMFPACK_DEFAULT_STRATEGY) ;
     if (strategy < UMFPACK_STRATEGY_AUTO
      || strategy > UMFPACK_STRATEGY_SYMMETRIC)
     {
@@ -153,7 +153,7 @@ GLOBAL void UMFPACK_report_control
     }
     else
     {
-	s = -alloc_init ;
+	s =(Int) -alloc_init ;
 	s = MAX (1, s) ;
 	PRINTF (("    "ID": initial allocation (in Units): "ID"\n",
 	(Int) INDEX (UMFPACK_ALLOC_INIT), s)) ;
@@ -163,7 +163,7 @@ GLOBAL void UMFPACK_report_control
     /* maximum iterative refinement steps */
     /* ---------------------------------------------------------------------- */
 
-    irstep = GET_CONTROL (UMFPACK_IRSTEP, UMFPACK_DEFAULT_IRSTEP) ;
+    irstep = (Int) GET_CONTROL (UMFPACK_IRSTEP, UMFPACK_DEFAULT_IRSTEP) ;
     irstep = MAX (0, irstep) ;
     PRINTF (("    "ID": max iterative refinement steps: "ID"\n",
 	(Int) INDEX (UMFPACK_IRSTEP), irstep)) ;
@@ -230,7 +230,7 @@ GLOBAL void UMFPACK_report_control
     /* scaling */
     /* ---------------------------------------------------------------------- */
 
-    scale = GET_CONTROL (UMFPACK_SCALE, UMFPACK_DEFAULT_SCALE) ;
+    scale = (Int) GET_CONTROL (UMFPACK_SCALE, UMFPACK_DEFAULT_SCALE) ;
     if (scale != UMFPACK_SCALE_NONE && scale != UMFPACK_SCALE_MAX)
     {
 	scale = UMFPACK_DEFAULT_SCALE ;
@@ -264,7 +264,7 @@ GLOBAL void UMFPACK_report_control
     }
     else
     {
-	s = -front_alloc_init ;
+	s = (Int) -front_alloc_init ;
 	s = MAX (1, s) ;
 	PRINTF (("    "ID": initial frontal matrix size (# of Entry's): "ID"\n",
 	(Int) INDEX (UMFPACK_FRONT_ALLOC_INIT), s)) ;
