+++ deal.II/contrib/umfpack/UMFPACK/Source/umfpack_report_triplet.c
@@ -33,7 +33,7 @@ GLOBAL Int UMFPACK_report_triplet
     Int prl, prl1, k, i, j, do_values ;
     Entry t ;
 
-    prl = GET_CONTROL (UMFPACK_PRL, UMFPACK_DEFAULT_PRL) ;
+    prl = (Int) GET_CONTROL (UMFPACK_PRL, UMFPACK_DEFAULT_PRL) ;
 
     if (prl <= 2)
     {
